refactor: per review
