+++ src/query/src/datafusion.rs
@@ -66,7 +66,6 @@ use crate::metrics::{
     OnDone, QUERY_STAGE_ELAPSED, maybe_attach_region_watermark_metrics,
     should_collect_region_watermark_from_query_ctx,
 };
-use crate::physical_wrapper::PhysicalPlanWrappersRef;
 use crate::planner::{DfLogicalPlanner, LogicalPlanner};
 use crate::query_engine::{DescribeResult, QueryEngineContext, QueryEngineState};
 use crate::{QueryEngine, metrics};
@@ -105,11 +104,6 @@ impl DatafusionQueryEngine {
 
         // `create_physical_plan` will optimize logical plan internally
         let physical_plan = self.create_physical_plan(&mut ctx, &plan).await?;
-        let physical_plan = if let Some(wrappers) = self.plugins.get::<PhysicalPlanWrappersRef>() {
-            wrappers.wrap(physical_plan, query_ctx, ctx.state().config_options())
-        } else {
-            physical_plan
-        };
         let physical_plan = self.optimize_physical_plan(&mut ctx, physical_plan)?;
 
         let stream = self.execute_stream(&ctx, &physical_plan)?;
